online_cinema/postgres_to_elastic/requirements.txt
python-dotenv==0.19.2
psycopg2-binary==2.9.3
pydantic==1.9.0
backoff==1.11.1
elasticsearch==7.14.1